Use exp package, don't use generics, split into separate file
