+++ go.mod
@@ -9,6 +9,7 @@ require (
 	github.com/nelkinda/health-go v0.0.1
 	github.com/oklog/run v1.1.0
 	github.com/prometheus/client_golang v1.20.6-0.20250117112434-e0800f53b498
+	github.com/prometheus/client_golang/exp v0.0.0-20250131124507-433de6814d4b
 	github.com/prometheus/client_model v0.6.1
 	github.com/prometheus/common v0.62.0
 	github.com/prometheus/prometheus v0.53.1
@@ -32,5 +33,3 @@ require (
 	google.golang.org/protobuf v1.36.1 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/prometheus/client_golang => ../client_golang
